requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
gunicorn==23.0.0
flask-cors==6.0.2
flask-debugtoolbar==0.16.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
python-dotenv==1.1.1
requests==2.31.0
pyodbc==5.3.0
Werkzeug==3.1.6
moviepy==1.0.3
numpy==1.26.4
pillow==10.2.0
imageio==2.34.0
